fix an ABareSliceMatrix issue, and a couple of size_t

+++ fem/intrule.cpp
@@ -225,9 +225,9 @@ namespace ngfem
   }
   */
 
-  IntegrationRule :: IntegrationRule (int asize, double (*pts)[3], double * weights)
+  IntegrationRule :: IntegrationRule (size_t asize, double (*pts)[3], double * weights)
   {
-    for (int i = 0; i < asize; i++)
+    for (size_t i = 0; i < asize; i++)
       {
 	IntegrationPoint ip(pts[i], weights[i]);
 	ip.SetNr(i);
@@ -238,7 +238,7 @@ namespace ngfem
 
   ostream & operator<< (ostream & ost, const IntegrationRule & ir)
   {
-    for (int i = 0; i < ir.GetNIP(); i++)
+    for (size_t i = 0; i < ir.GetNIP(); i++)
       ost << ir[i] << endl;
     return ost;
   }
